First commited as before_326_rev5

+++ org.eclipse.egit.ui/src/org/eclipse/egit/ui/internal/clone/GitCloneWizard.java
@@ -1,8 +1,9 @@
-/*******************************************************************************
+﻿/*******************************************************************************
  * Copyright (C) 2008, Roger C. Soares <rogersoares@intelinet.com.br>
  * Copyright (C) 2008, Shawn O. Pearce <spearce@spearce.org>
  * Copyright (C) 2008, Marek Zawirski <marek.zawirski@gmail.com>
  * Copyright (C) 2008, Robin Rosenberg <robin.rosenberg@dewire.com>
+ * Copyright (C) 2010, Mathias Kinzler <mathias.kinzler@sap.com>
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -14,6 +15,7 @@ package org.eclipse.egit.ui.internal.clone;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Collection;
+import java.util.Collections;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
@@ -21,49 +23,55 @@ import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.egit.core.op.CloneOperation;
 import org.eclipse.egit.ui.Activator;
+import org.eclipse.egit.ui.RepositoryUtil;
 import org.eclipse.egit.ui.UIIcons;
 import org.eclipse.egit.ui.UIText;
 import org.eclipse.egit.ui.internal.components.RepositorySelectionPage;
 import org.eclipse.jface.dialogs.ErrorDialog;
 import org.eclipse.jface.dialogs.MessageDialog;
-import org.eclipse.jface.viewers.IStructuredSelection;
+import org.eclipse.jface.operation.IRunnableWithProgress;
 import org.eclipse.jface.wizard.Wizard;
 import org.eclipse.jgit.lib.Ref;
 import org.eclipse.jgit.transport.URIish;
 import org.eclipse.osgi.util.NLS;
-import org.eclipse.ui.IImportWizard;
-import org.eclipse.ui.IWorkbench;
 
 /**
  * Import Git Repository Wizard. A front end to a git clone operation.
  */
-public class GitCloneWizard extends Wizard implements IImportWizard {
+public class GitCloneWizard extends Wizard {
 	private RepositorySelectionPage cloneSource;
 
 	private SourceBranchPage validSource;
 
 	private CloneDestinationPage cloneDestination;
 
-	private GitProjectsImportPage importProject;
+	String alreadyClonedInto;
 
-	public void init(IWorkbench arg0, IStructuredSelection arg1) {
+	/**
+	 * The default constructor
+	 */
+	public GitCloneWizard() {
 		setWindowTitle(UIText.GitCloneWizard_title);
 		setDefaultPageImageDescriptor(UIIcons.WIZBAN_IMPORT_REPO);
 		setNeedsProgressMonitor(true);
-		cloneSource = new RepositorySelectionPage(true);
-		validSource = new SourceBranchPage(cloneSource);
-		cloneDestination = new CloneDestinationPage(cloneSource, validSource);
-		importProject = new GitProjectsImportPage() {
+		cloneSource = new RepositorySelectionPage(true, null);
+		validSource = new SourceBranchPage() {
+
 			@Override
 			public void setVisible(boolean visible) {
-				if (visible) {
-					if (cloneDestination.alreadyClonedInto == null) {
-						if (performClone(false))
-							cloneDestination.alreadyClonedInto = cloneDestination
-									.getDestinationFile().getAbsolutePath();
-					}
-					setProjectsList(cloneDestination.alreadyClonedInto);
-				}
+				if (visible)
+					setSelection(cloneSource.getSelection());
+				super.setVisible(visible);
+			}
+
+		};
+		cloneDestination = new CloneDestinationPage() {
+			@Override
+			public void setVisible(boolean visible) {
+				if (visible)
+					setSelection(cloneSource.getSelection(), validSource
+							.getAvailableBranches(), validSource
+							.getSelectedBranches(), validSource.getHEAD());
 				super.setVisible(visible);
 			}
 		};
@@ -71,26 +79,30 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 
 	@Override
 	public boolean performCancel() {
-		if (cloneDestination.alreadyClonedInto != null) {
-			if (MessageDialog
-					.openQuestion(getShell(), "Aborting clone.",
-							"A complete clone was already made. Do you want to delete it?")) {
-				deleteRecursively(new File(cloneDestination.alreadyClonedInto));
+		if (alreadyClonedInto != null) {
+			File test = new File(alreadyClonedInto);
+			if (test.exists()
+					&& MessageDialog.openQuestion(getShell(),
+							UIText.GitCloneWizard_abortingCloneTitle,
+							UIText.GitCloneWizard_abortingCloneMsg)) {
+				deleteRecursively(new File(alreadyClonedInto));
 			}
 		}
 		return true;
 	}
 
 	private void deleteRecursively(File f) {
-		for (File i : f.listFiles()) {
-			if (i.isDirectory()) {
-				deleteRecursively(i);
-			} else {
-				if (!i.delete()) {
-					i.deleteOnExit();
+		File[] children = f.listFiles();
+		if (children != null)
+			for (File i : children) {
+				if (i.isDirectory()) {
+					deleteRecursively(i);
+				} else {
+					if (!i.delete()) {
+						i.deleteOnExit();
+					}
 				}
 			}
-		}
 		if (!f.delete())
 			f.deleteOnExit();
 	}
@@ -100,33 +112,41 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 		addPage(cloneSource);
 		addPage(validSource);
 		addPage(cloneDestination);
-		addPage(importProject);
 	}
 
 	@Override
 	public boolean canFinish() {
-		return cloneDestination.isPageComplete()
-				&& !cloneDestination.showImportWizard.getSelection()
-				|| importProject.isPageComplete();
+		return cloneDestination.isPageComplete();
 	}
 
 	@Override
 	public boolean performFinish() {
-		if (!cloneDestination.showImportWizard.getSelection())
-			return performClone(true);
-		return importProject.createProjects();
+		try {
+			return performClone(false);
+		} finally {
+			setWindowTitle(UIText.GitCloneWizard_title);
+		}
 	}
 
 	boolean performClone(boolean background) {
 		final URIish uri = cloneSource.getSelection().getURI();
-		final boolean allSelected = validSource.isAllSelected();
-		final Collection<Ref> selectedBranches = validSource
-				.getSelectedBranches();
+		setWindowTitle(NLS.bind(UIText.GitCloneWizard_jobName, uri.toString()));
+		final boolean allSelected;
+		final Collection<Ref> selectedBranches;
+		if (validSource.isSourceRepoEmpty()) {
+			// fetch all branches of empty repo
+			allSelected = true;
+			selectedBranches = Collections.emptyList();
+		} else {
+			allSelected = validSource.isAllSelected();
+			selectedBranches = validSource.getSelectedBranches();
+		}
 		final File workdir = cloneDestination.getDestinationFile();
 		final String branch = cloneDestination.getInitialBranch();
 		final String remoteName = cloneDestination.getRemote();
 
 		workdir.mkdirs();
+
 		if (!workdir.isDirectory()) {
 			final String errorMessage = NLS.bind(
 					UIText.GitCloneWizard_errorCannotCreate, workdir.getPath());
@@ -139,7 +159,12 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 
 		final CloneOperation op = new CloneOperation(uri, allSelected,
 				selectedBranches, workdir, branch, remoteName);
-		importProject.setGitDir(op.getGitDir());
+
+		alreadyClonedInto = workdir.getPath();
+
+		final RepositoryUtil config = Activator.getDefault()
+				.getRepositoryUtil();
+
 		if (background) {
 			final Job job = new Job(NLS.bind(UIText.GitCloneWizard_jobName, uri
 					.toString())) {
@@ -147,6 +172,8 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 				protected IStatus run(final IProgressMonitor monitor) {
 					try {
 						op.run(monitor);
+						cloneSource.saveUriInPrefs();
+						config.addConfiguredRepository(op.getGitDir());
 						return Status.OK_STATUS;
 					} catch (InterruptedException e) {
 						return Status.CANCEL_STATUS;
@@ -156,6 +183,7 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 								.getPluginId(), 0, thr.getMessage(), thr);
 					}
 				}
+
 			};
 			job.setUser(true);
 			job.schedule();
@@ -164,7 +192,18 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 			try {
 				// Perform clone in ModalContext thread with progress
 				// reporting on the wizard.
-				getContainer().run(true, true, op);
+				getContainer().run(true, true, new IRunnableWithProgress() {
+					public void run(IProgressMonitor monitor)
+							throws InvocationTargetException,
+							InterruptedException {
+						op.run(monitor);
+						if (monitor.isCanceled())
+							throw new InterruptedException();
+					}
+				});
+
+				cloneSource.saveUriInPrefs();
+				config.addConfiguredRepository(op.getGitDir());
 				return true;
 			} catch (InterruptedException e) {
 				MessageDialog.openInformation(getShell(),
@@ -172,9 +211,8 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 						UIText.GitCloneWizard_CloneCanceledMessage);
 				return false;
 			} catch (Exception e) {
-				Activator.logError(UIText.GitCloneWizard_CloneFailedHeading, e);
-				MessageDialog.openError(getShell(),
-						UIText.GitCloneWizard_CloneFailedHeading, e.toString());
+				Activator.handleError(UIText.GitCloneWizard_CloneFailedHeading,
+						e, true);
 				return false;
 			}
 		}
